changed the code for page.js,added code for amazonscrape.js and installed puppeteer package
