+++ package.json
@@ -21,7 +21,9 @@
     "@types/node": "^20.16.5",
     "@types/react": "^18.3.8",
     "ajv": "^8.17.1",
+    "axios": "^1.7.9",
     "chart.js": "^4.4.7",
+    "cheerio": "^1.0.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cors": "^2.8.5",
@@ -38,6 +40,8 @@
     "node-fetch": "^3.3.2",
     "node-schedule": "^2.1.1",
     "npm": "^10.8.3",
+    "playwright": "^1.49.1",
+    "proxycrawl": "^2.5.0",
     "puppeteer": "^23.11.1",
     "puppeteer-extra": "^3.3.6",
     "puppeteer-extra-plugin-stealth": "^2.11.2",
