--- src/app/shop/data/bigbasket-products.json
@@ -1,1 +1,0 @@
-[]
