--- src/app/shop/data/myntra-products.json
@@ -1,1 +1,0 @@
-[]
