--- src/app/shop/data/snapdeal-products.json
@@ -1,1 +1,0 @@
-[]
