+++ src/app/shop/page.js
@@ -1,11 +1,10 @@
 import React from "react";
 import { Star } from "lucide-react";
-import productsData from "./amazon-products.json";
+import productsData from "./data/amazon-products.json"; // Ensure this path is correct
 
 const ShoppingTracker = () => {
   const RatingStars = ({ rating }) => {
-    if (rating === "N/A")
-      return <span className="text-gray-400">No Rating</span>;
+    if (rating === "N/A" || !rating) return <span className="text-gray-400">No Rating</span>;
 
     const numRating = parseFloat(rating);
     const fullStars = Math.floor(numRating);
@@ -31,40 +30,43 @@ const ShoppingTracker = () => {
 
   return (
     <div className="container mx-auto px-4 py-8">
-      <h1 className="text-3xl font-bold text-center mb-8">
-        Sustainable Products
-      </h1>
+      <h1 className="text-3xl font-bold text-center mb-8">Sustainable Products</h1>
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-6">
-        {productsData.map((product, index) => (
-          <div
-            key={index}
-            className="bg-white rounded-lg shadow-lg overflow-hidden hover:shadow-xl transition-shadow duration-300"
-          >
-            <div className="relative pt-[100%]">
-              <img
-                src={product.imageUrl || "/api/placeholder/400/320"}
-                alt="Product placeholder"
-                className="absolute top-0 left-0 w-full h-full object-cover"
-              />
-            </div>
-            <div className="p-4">
-              <div className="flex justify-between items-start mb-2">
-                <span className="text-xl font-bold text-gray-800">
-                  {product.price}
-                </span>
-                <RatingStars rating={product.rating} />
+        {productsData && productsData.length > 0 ? (
+          productsData.map((product, index) => (
+            <div
+              key={index}
+              className="bg-white rounded-lg shadow-lg overflow-hidden hover:shadow-xl transition-shadow duration-300"
+            >
+              <div className="relative pt-[100%]">
+                <img
+                  src={product.imageUrl || "/api/placeholder/400/320"}
+                  alt={product.title || "Product"}
+                  className="absolute top-0 left-0 w-full h-full object-cover"
+                />
               </div>
-              <a
-                href={product.productUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mt-4 block w-full text-center bg-green-600 text-white py-2 rounded-md hover:bg-green-700 transition-colors duration-300"
-              >
-                View Product
-              </a>
+              <div className="p-4">
+                <div className="flex justify-between items-start mb-2">
+                  <span className="text-xl font-bold text-gray-800">{product.price}</span>
+                  <RatingStars rating={product.rating} />
+                </div>
+                <span className="text-sm text-gray-600">
+                  {product.productUrl && product.productUrl.includes("amazon") ? "Amazon" : "Flipkart"}
+                </span>
+                <a
+                  href={product.productUrl || "#"}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="mt-4 block w-full text-center bg-green-600 text-white py-2 rounded-md hover:bg-green-700 transition-colors duration-300"
+                >
+                  View Product
+                </a>
+              </div>
             </div>
-          </div>
-        ))}
+          ))
+        ) : (
+          <p>No products available.</p>
+        )}
       </div>
     </div>
   );
